Non-complete more-comprehensive chat.
This makes a lot of changes to the chat system, but it is not working
as of this moment. Changes include adding the newchat_* files in the
base module as well as a few other files and adding lots of images to
make the chat UI smaller and more self-explanatory.