Got chat system working preliminarially
Got Chat Alpha 1 (Out of a couple hundred?) working!
There are quite a few bugs with this system, but it encompasses most of the
functionality that will be added, except for the necessary replication to
the client.  I have to go through and clean up the code and fix a few bugs
that allow invalid NULLs to be inserted into the tables.