David: Update update_descriptives_new.qy
I questioned a part of block of code by putting a comment on it.  Is there any exception handling? From thinking about that, I have an idea of what to do with the data we have: 
Imagine if there is a huge set of data, then people is categorized into different boxes which represents if they are supposed to pay annually, twice, or even quarterly. Then from there, use exception handling for each person/category/box to go through each block of tests. Put each of the boxes into the exception handling, and catch all of it like if it is supposed to pay only once. Like when the box passes an exception of handling one its green means good (when it passes the test of paying only once), but also detect (put red) when the pay is increased, more than once, etc. So it is like a box going through a series of checks. Then once it is finished, display the results (both red and green) so it could be further examined and understood. But is it better if it is with exception handling or something else? 

I also found out that there are lines of code around 500-750 where there are series of checks. Are these only for the history data? Can it (the checks) be incorporated with the data that we are going to analyze?

Thanks